+++ src/components/login/Splash/index.jsx
