+++ src/components/login/Splash/index.style.js
