+++ src/components/product/AddProduct/index.jsx
@@ -1,32 +1,26 @@
-import React, {useState} from 'react'
-import { useDispatch, useSelector } from 'react-redux';
+import React, { useState } from 'react';
+import { useDispatch } from 'react-redux';
 import { useHistory } from 'react-router-dom';
-import { productAction } from '../../../redux/actions/productAcition';
-import { SaveBtn } from '../../common/Buttons';
+import { productAction } from '../../../redux/actions/productAcition.js';
+import { SaveBtn } from '../../common/button/index.style';
+import { ProductLink, ProductName, ProductPrice, TextLabel } from '../../common/TextAciveInput.jsx';
 import { ArrowLeftLink } from '../../common/nav';
-import {TopNavRowBox} from '../../common/nav/index.style.jsx'
-import { AddProductImgInput, AddProductLabel, AddProductSpan, ProductBox, ProductForm, ProductNameLabel } from './index.style';
-import { ProductLink, ProductName, ProductPrice, TextLabel } from '../../common/TextAciveInput';
+import {TopNavRowBox} from '../../common/nav/index.style.jsx';
+import { AddProductImgInput, AddProductLabel, AddProductSpan, ProductBox, ProductForm, ProductNameLabel } from './index.style.js';
 
-function ProductModification() {
-    const product_id = useSelector(state=>state.product.product_id);
-    const uitemName = useSelector(state=>state.product.itemName);
-    const uprice = useSelector(state=>state.product.price);
-    const ulink = useSelector(state=>state.product.link);
-    const uitemImage = useSelector(state=>state.product.itemImage);
-
-    const [itemName, setItemName] = useState(uitemName);
-    const [price, setPrice] = useState(uprice);
-    const [isPrice, setIsPrice] = useState(uprice);
-    const [link, setLink] = useState(ulink);
-    const [itemImage, setItemImage] = useState(uitemImage);
-    const [isActive, setisActive] = useState(false);
+function AddProduct() {
+    const [itemName, setItemName] = useState('');
+    const [price, setPrice] = useState('');
+    const [isPrice, setIsPrice] = useState('');
+    const [link, setLink] = useState('');
+    const [itemImage, setItemImage] = useState('');
+    const [previewImage, setPreviewImage] = useState('');
+    const [isActive, setisActive] = useState(true);
     const history = useHistory();
     const dispatch = useDispatch();
 
     const saveActive = () => {
-        return itemName.length>1&&itemName.length<16&&link.length>0&&itemImage.length>0
-        // &&isPrice.length>0&&link.length>0&&itemImage.length>0
+        return itemName.length>1&&itemName.length<16&&isPrice.length>0&&link.length>0&&previewImage.length>0
             ? setisActive(false)
             : setisActive(true);
     }
@@ -34,7 +28,7 @@ function ProductModification() {
     const onSubmitHandler = (event) => {
         event.preventDefault();
         console.log('onSubmitHandler');
-        dispatch(productAction.productModification(itemName, price, link, itemImage, product_id));
+        dispatch(productAction.addProduct(itemName, price, link, itemImage));
         history.push('/myprofile');
     }
 
@@ -46,6 +40,8 @@ function ProductModification() {
     }
 
     const onChangeProductImg = (event) => {
+        setPreviewImage(URL.createObjectURL(event.target.files[0]));
+
         let reader = new FileReader();
         reader.readAsDataURL(event.target.files[0]);
         reader.onload = (event) => {
@@ -53,7 +49,6 @@ function ProductModification() {
             setItemImage(readerUrl);
         }
     };
-
     return (
         <ProductForm onSubmit={onSubmitHandler}>
             <TopNavRowBox>
@@ -77,4 +72,4 @@ function ProductModification() {
     )
 }
 
-export default ProductModification
+export default AddProduct;
