+++ src/pages/AddProductPage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import AddProduct from '../components/';
+import AddProduct from '../components/product/AddProduct';
 
 function AddProductPage() {
     return (
